+++ Task2/Lab1_3/3DShape/3DShape.h
@@ -94,7 +94,7 @@ public:
 	//--------------------------------------------
 
 	std::vector<SVertexP3NT2>		GetVertexes() const override;
-	glm::vec3						GetVertexPosition(uint index) const override;
+	glm::vec3						GetVertexPosition(size_t index) const override;
 
 //////////////////////////////////////////////////////////////////////
 // Data
